+++ kernel/shell/apps/uname.c
@@ -1,6 +1,7 @@
 #include "uname.h"
 #include "console.h"
+#include "nexus_version.h"
 
 void uname_run(void) {
-    console_print("NexusOS kernel 0.3-experimental x86_64\n");
+    console_print("NexusOS kernel " NEXUS_VERSION_STRING " x86_64\n");
 }